backend/pyproject.toml
[project]
name = "codette-backend"
version = "0.1.0"
description = "Codette backend API and worker"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.30",
    "psycopg[binary]>=3.1.18",
    "alembic>=1.13.1",
    "pydantic-settings>=2.2.1",
    "google-cloud-storage>=2.16.0",
    "google-cloud-run>=0.10.0",
    "tenacity>=8.2.3",
    "docker>=7.0.0",
    "PyJWT>=2.8.0",
    "argon2-cffi>=23.1.0",
    "pydantic[email]>=2.0.0",
    "ipython>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.4",
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ipdb"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.coverage.run]
source = ["app"]

[tool.coverage.report]
show_missing = true
skip_empty = true

[tool.ruff]
target-version = "py312"
line-length = 88
extend-exclude = ["migrations/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["app"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
